replace commentable with story everywhere, consolidate controllers and routes, move views
