+++ app/models/comment.rb
@@ -4,21 +4,21 @@
 #
 # Table name: comments
 #
-#  id               :integer          not null, primary key
-#  user_id          :integer          not null
-#  commentable_type :string           not null
-#  commentable_id   :integer          not null
-#  parent_id        :integer
-#  created_at       :datetime         not null
-#  updated_at       :datetime         not null
-#  level            :integer          default(1)
+#  id              :integer          not null, primary key
+#  user_id         :integer          not null
+#  parent_id       :integer
+#  created_at      :datetime         not null
+#  updated_at      :datetime         not null
+#  root_comment_id :integer
+#  level           :integer          default(0)
+#  story_id        :integer          not null
 #
 class Comment < ApplicationRecord
   belongs_to :user
-  belongs_to :commentable, polymorphic: true
+  belongs_to :story, inverse_of: :comments
   belongs_to :parent, optional: true, class_name: 'Comment', inverse_of: :comments
-  has_one :pin, class_name: 'Pin', dependent: :destroy
   has_many :comments, foreign_key: :parent_id, dependent: :destroy, inverse_of: :parent
+  has_one :pin, class_name: 'Pin', dependent: :destroy
 
   has_rich_text :content
   scope :joins_content, lambda {
