+++ app/models/story.rb
@@ -15,7 +15,7 @@ class Story < ApplicationRecord
   belongs_to :user
 
   has_rich_text :content
-  has_many :comments, as: :commentable
+  has_many :comments, dependent: :destroy, inverse_of: :story
   has_many :pins, dependent: :destroy, inverse_of: :story
 
   validates :title, presence: true
