--- app/views/comments/_form.html.erb
@@ -1,16 +1,0 @@
-<% model = [local_assigns[:commentable], comment].compact %>
-
-<%= form_with model: model,
-  id: dom_id_for_records(*model),
-  data: local_assigns[:data],
-  class: ["my-4"] + Array.wrap(local_assigns[:class]) do |form|
-%>
-  <%= form.rich_text_area :content, id: "#{dom_id_for_records(*model)}_content", autocomplete: "content", class: "form-control", rows: 4 %>
-  <%= render partial: "shared/errors", locals: { resource: form.object, attribute: :content } %>
-
-  <div class="d-flex align-items-center mt-3 mb-3">
-    <%= form.submit t('save'), class: "btn btn-m btn-primary me-3" %>
-    <!-- TODO: after cancel, the stimulus doesn't reload, so the hideable-buttons disappear -->
-    <%= link_to t('cancel'), comment, class: "btn btn-link", data: { turbo_method: :get, turbo_confirm: t('are_you_sure_changes_unsaved') } if form.object.persisted? %>
-  </div>
-<% end %>
