+++ app/views/pins/_pin.html.erb
@@ -1,7 +1,7 @@
 <%= turbo_stream_from dom_id(pin) %>
 <%= tag.div id: dom_id(pin) do %>
   <div class="d-flex flex-column w-100">
-    <%= link_to story_comment_path(pin.story.id, pin.comment.id), class: "pin text-decoration-none text-reset mb-2 ps-2", data: { controller: 'turbo', action: 'click->turbo#getTurboStream' } do %>
+    <%= link_to story_comment_path(story_id: pin.story_id, id: pin.comment_id), class: "pin text-decoration-none text-reset mb-2 ps-2", data: { controller: 'turbo', action: 'click->turbo#getTurboStream' } do %>
       <div class="py-2">
         <span id="<%= dom_id(pin) %>_content"><%= pin.comment.content %></span>
       </div>
