+++ app/views/pins/button/_pin.html.erb
@@ -1,3 +1,3 @@
-<%= tag.div id: "#{dom_id(comment)}_pin", data: { controller: "turbo-frame-visibility", turbo_frame_visibility_target: "hideableTurboFrame", username_value: "#{comment.commentable.user.username}" }, class: "d-none" do %>
-  <%= button_to 'Pin', pins_path(pin: { story_id: comment.commentable.id, comment_id: comment.id }), data: { action: 'click->pin-comment#toggleClass' }, class: "btn btn-link" %>
+<%= tag.div id: "#{dom_id(comment)}_pin", data: { controller: "turbo-frame-visibility", turbo_frame_visibility_target: "hideableTurboFrame", username_value: "#{comment.story.user.username}" }, class: "d-none" do %>
+  <%= button_to 'Pin', pins_path(pin: { story_id: comment.story_id, comment_id: comment.id }), data: { action: 'click->pin-comment#toggleClass' }, class: "btn btn-link" %>
 <% end %>
