+++ app/views/pins/button/_unpin.html.erb
@@ -1,3 +1,3 @@
-<%= tag.div id: "#{dom_id(comment)}_pin", data: { controller: "turbo-frame-visibility", turbo_frame_visibility_target: "hideableTurboFrame", username_value: "#{comment.commentable.user.username}" }, class: "d-none" do %>
+<%= tag.div id: "#{dom_id(comment)}_pin", data: { controller: "turbo-frame-visibility", turbo_frame_visibility_target: "hideableTurboFrame", username_value: "#{comment.story.user.username}" }, class: "d-none" do %>
   <%= button_to 'Unpin', pin_path(comment.pin), data: { action: 'click->pin-comment#toggleClass' }, class: "btn btn-link", method: :delete %>
 <% end %>
