+++ app/views/stories/comments/_comment_with_replies.html.erb
@@ -1,7 +1,7 @@
 <%= turbo_stream_from "#{dom_id(comment)}_comments" %>
 
 <div class="<%= 'ps-3 border-left' unless comment.parent.nil? || defined?(is_root_comment) %> mt-3" id="<%= dom_id(comment) %>_with_comments" data-controller="comment-reply pin-comment">
-  <%= render partial: 'comments/comment', locals: { comment: comment } %>
+  <%= render partial: 'stories/comments/comment', locals: { comment: comment } %>
 
   <div class="text-sm mt-1 mb-3 d-flex">
     <%= link_to t('reply'), "#", id: "#{dom_id(comment)}_reply_button", class: "btn btn-link", data: { action: "click->comment-reply#toggle" } %>
@@ -9,17 +9,17 @@
   </div>
 
   <div id="<%= dom_id(comment) %>_comments">
-    <%= render partial: "comments/form", locals: { commentable: comment, comment: Comment.new, data: { comment_reply_target: "form" }, class: "d-none" } %>
+    <%= render partial: "stories/comments/form", locals: { target: comment, comment: comment.story.comments.new, data: { comment_reply_target: "form" }, class: "d-none" } %>
     <% if comment.level.to_i != 0 && comment.level.to_i % 10 == 0 && comment.comments.any? && !defined?(is_root_comment) %>
       <div class="ms-2 p-1">
         <%= link_to(
           "Continue this thread",
-          story_comment_path(comment.commentable.id, comment.id),
+          story_comment_path(story_id: comment.story_id, id: comment.id),
           data: { controller: 'turbo', action: 'click->turbo#getTurboStream' }
         ) %>
       </div>
     <% else %>
-      <%= render partial: "comments/comment_with_replies", collection: comment.comments, as: :comment %>
+      <%= render partial: "stories/comments/comment_with_replies", collection: comment.comments, as: :comment %>
       <!-- TODO: comment validation errors are not rendered for nested comments -->
     <% end %>
   </div>
