+++ app/views/stories/comments/destroy.turbo_stream.erb
