+++ app/views/stories/comments/update.turbo_stream.erb
@@ -1,4 +1,4 @@
 <%= turbo_stream.replace dom_id(@comment) do %>
-  <%= render partial: 'comments/comment', locals: { comment: @comment } %>
+  <%= render partial: 'stories/comments/comment', locals: { comment: @comment } %>
 <% end %>
 <%= turbo_stream.prepend "flash", partial: "shared/flash" %>
